Struct embre::aes::aes_resource::AESResource

pub struct AESResource<'a, T, C: AESCrypter<'a> = DefaultAesCrypter> { /* private fields */ }

Implementations

impl<'a, T> AESResource<'a, T>

pub fn new( cipher_text: &'a [u8], key: &'a [u8], iv: Option<&'a [u8]> ) -> AESResource<'a, T>

pub fn to_plaintext_data(&self) -> Result<Vec<u8>>

Returns a Vec<u8> with the decrypted data.

pub fn get_encrypted_slice(&self) -> &'a [u8]

pub fn get_key(&self) -> &'a [u8]

pub fn get_iv(&self) -> Option<&'a [u8]>

impl<'a, T, C: AESCrypter<'a>> AESResource<'a, T, C>

pub fn new_from( cipher_text: &'a [u8], key: &'a [u8], iv: Option<&'a [u8]>, crypter: C ) -> AESResource<'a, T, C>

impl<'a> AESResource<'a, StringResource>

pub fn to_plaintext_string(&self) -> Result<String, FromUtf8Error>

This returns the original plaintext version of the string in a new String

pub fn to_plaintext_c_string(&self) -> Result<CString, NulError>

This returns the original plaintext version of the string in a new null terminated CString