Type Alias embre::aes::aes_string::AESString

pub type AESString<'a> = AESResource<'a, StringResource>;

Aliased Type

struct AESString<'a> { /* private fields */ }

Implementations

impl<'a> AESString<'a>

pub fn to_plaintext_string(&self) -> Result<String, FromUtf8Error>

This returns the original plaintext version of the string in a new String

pub fn to_plaintext_c_string(&self) -> Result<CString, NulError>

This returns the original plaintext version of the string in a new null terminated CString