Trait embre::embedded_resource::EmbeddedAES

pub trait EmbeddedAES: EmbeddedResource {
    // Provided methods
    fn get_str(&self, offsets: AESOffsets) -> AESString<'static> { ... }
    fn get_data(&self, offsets: AESOffsets) -> AESData<'static> { ... }
}

Provided Methods

fn get_str(&self, offsets: AESOffsets) -> AESString<'static>

Returns the encrypted data from the provided offsets, as an AESString.

fn get_data(&self, offsets: AESOffsets) -> AESData<'static>

Returns the encrypted data from the provided offsets, as AESData.